Unify jump links into sticky element
The page can be quite long, with up to 25 tweets shown at a time.
Getting back to the top from the middle of the page on mobile requires a
lot of scrolling. This makes it easier by keeping the top jump link in
view when you scroll down. The bottom jump link is now redundant and
therefore removed